Make RuleSet for window reorder and merge rules
Because MergeAdjacentWindows and SwapAdjacentWindowsBySpecifications are
very much related, and a future commit will introduce more commonality,
move them together into a RuleSet.  Leave them as public so that the
existing unit testing of them as individual rules is still possible.
